added basic mouse support
In a terminal which supports mouse tracking (such as TERM=xterm-1002),
right-click-and-drag marks text for editing.

In a terminal which does not support mouse tracking (such as TERM=xterm),
left-click-and-drag marks text for editing (but does not show highlighting
until drag is over).

In all terminals, left click positions cursor.

	modified:   editor.rb