automatic folding/hiding based on start & end patterns
Storing up patterins in a history.
Should be backwards compatible with old history file, but old code
not compatible with new history file.

One issue: line numbers all correspond to the folded text,
not the actual text.  This is because, to the editor, the folded
text *is* the text.  To change all line number oriented code to
compute actual line numbers would make things really messy.

	modified:   editor.rb