bugfix: empty file should be stored as [""] not []
Philosophically, what is an empty file?
Is it a file with no lines?  Or a file with a single empty line.
Equivalently, is the last newline of a file followed by a blank line
or no line?

Practically, it does not matter which we choose, as long as we are consistent.
I have decided to choose the former (blank line follows last newline), and
this bugfix makes that consistent throughout the code.

	modified:   editor.rb