bugfix: editor was crashing if we hadn't saved recently :)
It was trying to compared current state to last saved state, which had
been chopped off to keep undo list from being too long.  I changed from
a linke list to an array.  This presents some challenges with keeping
the saved states indices updated when we prune the history list. But it
still seems cleaner than the linked list.

	modified:   editor.rb