bugfix: don't depend on terminal non-linewrapping
Some terminals, when line-wrapping is disabled, print the remainder of
the line on the last column of the screen.  This commit make sure to
only print the exact number of characters we need.

Needs to be tested. Does it slow things down too much.  Unlike the
colfeed stuf, this must be run on every printed line.

	modified:   editor.rb