Better handling of character encodings
This commit involves three things:

1. Allow a user to force ascii-only display.  This is useful if we are
using ruby 1.8 (no builtin unicode support) and we have strange
characters.  Multibyte characters especially cause problems, because
the cursor is off by one.  This option (default ctrl-D/d) will display
any non-ascii bytes as a tilde.

2. Handle mixed character encodings.  If there are some unicode and
some ascii strings in a buffer, the lines cannot be concatenated into
one big output string (for file writing).  If this happens, the user is
presented with the option of forcing unicode assumptions on all the
string. This is probably perfectly safe, but I give the user the
heads-up just in case.

3. Force reading/writing of UTF-8 if available.  If we are in ruby >
1.9, then files will be read/written in UTF-8 all the time.  This
should be fine, except maybe for writing japanese characters.

	modified:   src/editor.rb
	modified:   src/filebuffer.rb
	modified:   src/keymap.rb