feature: search for startup scripts in the current path.
Look for startup scripts in $HOME/.jedi/, and then in current path. If
the current path includes $HOME, then only look between $HOME and the
working directory.

This allows a user to have local config files. For example, if there is
one project that requires certain settings, this allows you to make
those local settings.

Installation script has been modified to incorporate the changes.  In
particular, no startup script is needed anymore. One side effect is
that the executable can no longer be renamed, because we are looking
for .jedi/.  The -P flag turns off this searching, which returns the
old behavior.

	modified:   install.sh
	modified:   make_jedi.rb
	modified:   scripts/config.rb
	modified:   src/editor.rb