bugfix: undo/redo errors at the end of the file.
The issue was: if lines disappeared during and edit, then to try to
recover that change required first moving to the line that the change
occurred on. This was added as a feature some time ago, so the user
could see exactly what was happening durning and undo/redo.  In this
situation, because the cursor can never get to the edited line, the
change was inaccessible. This fix now handles that situation.

I also simplifed undo/redo a little bit.

	modified:   src/bufferhistory.rb
	modified:   src/filebuffer.rb