bugfix: fix problem with indentation facade
If a file with history is opened, and immediately we change the
indentaion facade, it does not apply to the current state (only to the
history).

Here's what was happening: the initial text state is not part of the
history, because when it is read in from a file, it hasn't changed
state from the previous version (probably).  Thus, if the current (from
file) state matches the "current" state in the buffer history, set the
current state to a clone of the buffer history current state.

	modified:   src/filebuffer.rb